refactor code

+++ data_model/model/llm.py
@@ -7,11 +7,11 @@ class TextOutput(BaseModel):
     text: str = Field("")
 
 class SpeedOutput(BaseModel):
-    tokens_per_second: int = Field(
-        default=0,
+    tokens_per_second: float = Field(
+        default=0.0,
         description="tokens per second"
     )
-    avg_time_per_iteration: int = Field(
-        default=0,
+    avg_time_per_iteration: float = Field(
+        default=0.0,
         description="avg time per iteration (seconds)"
     )
